# Conflicts:
#	events/ready.js
#	package-lock.json
#	package.json

+++ events/ready.js
@@ -57,16 +57,19 @@ const deleteExpiredSavedRollCommands = async () => {
   }
 }
 
-const removeWarningInteractivity = async (client, warning) => {
+const removeWarningInteractivity = async (warning) => {
   let channel
   try {
-    channel = await client.channels.fetch(warning[MESSAGES_COLUMNS.channel_id])
+    channel = await Client.getChannelById(warning[MESSAGES_COLUMNS.channel_id])
   } catch (error) {
     if (error && error.name === 'DiscordAPIError') {
       logger.warn(`Failed to fetch channel for a warning`, error)
     } else {
       logger.error(`Failed to fetch channel for a warning`, error)
     }
+    return
+  }
+  if (!channel) {
     return
   }
   let message
@@ -80,7 +83,9 @@ const removeWarningInteractivity = async (client, warning) => {
     }
     return
   }
-
+  if (!message) {
+    return
+  }
   try {
     await message.delete()
   } catch (error) {
@@ -112,7 +117,7 @@ const deleteExpiredWarningMessages = async () => {
         const channelId = warning[MESSAGES_COLUMNS.channel_id]
         const messageId = warning[MESSAGES_COLUMNS.message_id]
         Client.reactionsCache[channelId+ '_' + messageId] = null
-        removeWarningInteractivity(Client.client, warning)
+        removeWarningInteractivity(warning)
       })
     }
   } catch (error) {
@@ -174,6 +179,17 @@ module.exports = async (client) => {
   }
   else {
     logger.log(`No shard`)
+  }
+
+  try {
+    const channel = await Client.getChannelById(process.env.LOG_CHANNEL_ID)
+    if (channel) {
+      const mes = await Client.getMessageByIdAndChannelId(
+        channel.lastMessageId, channel.id)
+      let i = 0
+    }
+  } catch (err) {
+    logger.error(`Failed to get channel`, err)
   }
 
   await tryToSetActivity()
